Add files via upload
The idea behind the project is to help one of my friend to grow his newly started Business of Food Blogging handle on Instagram and wants to get famous. He wants to follow a lot of people so that he can get noticed quickly. So through automation using Selenium, I have decided to help him by creating an Instagram Bot